add restore not clear

+++ content/moduls.php
@@ -1,13 +1,17 @@
 <?php
-    $id_user = isset($_SESSION['NAME']) ? $_SESSION['NAME'] : '';
+    $id_user = isset($_SESSION['ID_USER']) ? $_SESSION['ID_USER'] : '';
     $id_role = isset($_SESSION['ID_ROLE']) ? $_SESSION['ID_ROLE'] : '';
     
     $rowStudent = mysqli_fetch_assoc(mysqli_query($config, "SELECT * FROM students WHERE id='$id_user'"));
     $id_major = $rowStudent['id_major'];
-    if ($id_role == 2) {
+    // print_r($id_major);
+    // die;
+    if ($id_role == 6) {
         $where = "WHERE md.id_major='$id_major'";
-    } elseif ($id_role == 1) {
+    } elseif ($id_role == 2) {
         $where = "WHERE md.id_instructor='$id_user'";
+    } else {
+        $where = "";
     }
     $query = mysqli_query($config,"SELECT mj.name major, i.name instructor, md.*  FROM moduls md
     LEFT JOIN majors mj ON md.id_major = mj.id
@@ -22,7 +26,7 @@
         <div class="card">
             <div class="card-body">
                 <h5 class="card-title">Data Modul</h5>
-                <?php if ($_SESSION['ID_ROLE'] == 1) { ?>    
+                <?php if (canAddModul($id_role)) { ?>    
                     <div class="mb-3" align="right">
                         <a href="?page=tambah-modul" class="btn btn-primary">+</a>
                     </div>
@@ -46,8 +50,10 @@
                                     <td><?php echo $data['instructor']; ?></td>
                                     <td><?php echo $data['major']; ?></td>
                                     <td>
-                                        <a href="?page=tambah-modul&edit=<?php echo $data['id']; ?>" class="btn btn-success">Edit</a>
-                                        <a onclick="return confirm('Are you sure?')" href="?page=tambah-modul&delete=<?php echo $data['id'] ?>" class="btn btn-danger">Delete</a>
+                                        <?php if ($id_role == 1) { ?>
+                                            <a href="?page=tambah-modul&edit=<?php echo $data['id']; ?>" class="btn btn-success">Edit</a>
+                                            <a onclick="return confirm('Are you sure?')" href="?page=tambah-modul&delete=<?php echo $data['id'] ?>" class="btn btn-danger">Delete</a>
+                                        <?php } ?>
                                     </td>
                                 </tr>
                             <?php } ?>
